feat(cli): forward JENTIC_API_BASE_URL into the container
When the host env sets JENTIC_API_BASE_URL, the CLI now forwards it
to the container with the same loopback rewrite the LLM URL passthrough
already does (host.docker.internal on macOS/Win, --network host on
Linux). Extract the shared loopback helpers into docker-host.ts so
llm-env.ts and score.ts both consume them.

Threat model unchanged: `docker run -e JENTIC_API_BASE_URL=…` against
the public image already bypassed the validator without the CLI's help.
The runner-side override has always been documented as a test hook.
This unblocks JS-side e2e coverage of the validator paths.
