+++ packages/cli/src/commands/score.ts
@@ -7,6 +7,7 @@ import { ExitCode } from '../exit-codes.ts';
 import { DEFAULT_FORMAT, Format } from '../format.ts';
 import { formatJson } from '../formatters/json.ts';
 import { formatPretty } from '../formatters/pretty.ts';
+import { needsDockerHostRewrite, rewriteUrlForContainer } from '../docker-host.ts';
 import { detectLlmEnv } from '../llm-env.ts';
 import { writeReport } from '../output.ts';
 import { ScorecardResult } from '../result.ts';
@@ -125,6 +126,18 @@ export async function runScore(input: string, options: ScoreOptions): Promise<nu
     forwardEnvVars = detection.forwardEnvVars;
     forwardEnvOverrides = detection.forwardEnvOverrides;
     needsHostNetwork = detection.needsHostNetwork;
+  }
+
+  const baseUrl = process.env['JENTIC_API_BASE_URL'];
+  if (baseUrl !== undefined && baseUrl !== '') {
+    forwardEnvVars.push('JENTIC_API_BASE_URL');
+    if (needsDockerHostRewrite(baseUrl)) {
+      needsHostNetwork = true;
+      const rewritten = rewriteUrlForContainer(baseUrl);
+      if (rewritten !== baseUrl) {
+        forwardEnvOverrides.set('JENTIC_API_BASE_URL', rewritten);
+      }
+    }
   }
 
   let stdinPayload: string | undefined;
